pyproject.toml
[tool.poetry]
name = "amplify_routing"
version = "0.1.0"
description = ""
authors = ["KO <keisuke.otaki.jp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.0"
matplotlib = "^3.5.1"
amplify = "^0.6.5"
jupyterlab = "^3.2.5"
ipython = "^7.31.0"
python-dotenv = "^0.19.2"
tqdm = "^4.62.3"
optuna = "^2.10.0"
streamlit = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
